requirements.txt
PySide2
requests
psutil
RPI.GPIO; sys_platform == 'linux'